point to @streamdal/protos, point examples to @streamdal/node-sdk, finish rename
