+++ src/streamdal.ts
@@ -6,11 +6,11 @@ import {
   OperationType,
   StandardResponse,
   TailResponse,
-} from "@streamdal/snitch-protos/protos/sp_common";
+} from "@streamdal/protos/protos/sp_common";
 import {
   IInternalClient,
   InternalClient,
-} from "@streamdal/snitch-protos/protos/sp_internal.client";
+} from "@streamdal/protos/protos/sp_internal.client";
 import { v4 as uuidv4 } from "uuid";
 
 import { addAudience, addAudiences } from "./internal/audience.js";
